# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for images",
        "title": "Imloader Server API",
        "contact": {},
        "version": "0.1"
    },
    "paths": {
        "/get/{id}": {
            "get": {
                "description": "return image by its id",
                "produces": [
                    "image/jpeg"
                ],
                "summary": "get image by its id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "image id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpext.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpext.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "upload jpg image into MySQL database (max allowed size - 50 mb)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Upload jpg image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "image to upload",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.UploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpext.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpext.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.UploadResponse": {
            "type": "object",
            "properties": {
                "imageToken": {
                    "type": "string"
                }
            }
        },
        "httpext.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}